Fix order of operations to step timer gems first
We previously stepped timer gems as the last operation in a turn; this
results in all kinds of strangeness, such as timer gems stepping
immediately on dropping and not being destroyed until a turn later than
they should. This is fixed by making timer gem stepping the first
operation to be performed after a turn is taken.